// packages/tsts/src/internal/ast/generated/encoder.ts
export const TargetAstNodeDataChildren = 0;
export const TargetAstNodeDataString = 1;
export const TargetAstNodeDataExtended = 2;

export interface TargetAstEncodedChild {
  readonly name: string;
  readonly present: boolean;
  readonly required: boolean;
  readonly raw: boolean;
  readonly node?: GoPtr<Node>;
  readonly nodes?: readonly GoPtr<Node>[];
  readonly list?: GoPtr<NodeList>;
}

export interface TargetAstNodeEncoding {
  readonly dataType: 0 | 1 | 2;
  readonly commonData: number;
  readonly children: readonly TargetAstEncodedChild[];
  readonly text?: string;
  readonly rawText?: string;
  readonly tokenFlags?: number;
  readonly extended: "literal" | "template" | "source-file" | "none";
}